tests/testdata/tree-construction/button.dat

#data
<button><p></button><button></button>
#errors
#document
| <html>
|   <head>
|   <body>
|     <button>
|       <p>
|     <button>
#data
<button><b><button>hei
#errors
#document
| <html>
|   <head>
|   <body>
|     <button>
|       <b>
|     <b>
|       <button>
|         "hei"
